Move lua unit tests to thier own file
